[Chore #160088425] Restructure files
