+++ src/Home/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Loader from '../../views/Loader';
+import Loader from '../views/Loader';
 
 import './index.scss';
