+++ src/Signup/components/Signup.css
@@ -88,7 +88,7 @@ button {
     padding-top: 20px;
     display: flex;
     justify-content: space-between;
-    margin-left: 350px;
+    margin-left: 300px;
 
 }
