+++ src/views/App/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../components/Header';
+import Header from '../../Header';
 import Main from '../Main';
 
 const App = () => (
