+++ tests/components/homepage.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, } from 'enzyme';
-import Home from '../../src/components/Home';
+import Home from '../../src/Home';
 
 describe('Home', () => {
   it('renders without crashing', () => {
