+++ tests/components/login.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, } from 'enzyme';
-import Login from '../../src/components/Login';
+import Login from '../../src/Login';
 
 describe('Login', () => {
   it('renders without crashing', () => {
